refactored project structure with a test directory
