+++ CHANGELOG.md
@@ -1,49 +1,39 @@
 # Changelog
 
-## [v0.1.2](https://github.com/ibbatta/readme-generator/compare/v0.1.1...v0.1.2) - 2019-08-28
+## Unreleased
 
 ### Merged
 
-- Refactory [`#36`](https://github.com/ibbatta/readme-generator/pull/36)
+- Refactory [`#36`](https://github.com/audreylopes/readme-generator/pull/36)
+- Packagejson fallback [`#35`](https://github.com/audreylopes/readme-generator/pull/35)
+- Jest test [`#34`](https://github.com/audreylopes/readme-generator/pull/34)
+- Change babel polyfill [`#32`](https://github.com/audreylopes/readme-generator/pull/32)
+- Added link to generated example in README [`#30`](https://github.com/audreylopes/readme-generator/pull/30)
+- Develop [`#28`](https://github.com/audreylopes/readme-generator/pull/28)
+- Flag options [`#27`](https://github.com/audreylopes/readme-generator/pull/27)
+- Webpack bundler [`#26`](https://github.com/audreylopes/readme-generator/pull/26)
+- Npm packaging [`#25`](https://github.com/audreylopes/readme-generator/pull/25)
+- Develop [`#24`](https://github.com/audreylopes/readme-generator/pull/24)
+- parse yaml file (FUNDING.yml) to autocomplete patreon questions [`#23`](https://github.com/audreylopes/readme-generator/pull/23)
+- Develop [`#19`](https://github.com/audreylopes/readme-generator/pull/19)
+- Update readme template [`#18`](https://github.com/audreylopes/readme-generator/pull/18)
+- Develop [`#17`](https://github.com/audreylopes/readme-generator/pull/17)
+- Align develop with master [`#14`](https://github.com/audreylopes/readme-generator/pull/14)
+- Refactor settings [`#13`](https://github.com/audreylopes/readme-generator/pull/13)
+- Bump lodash from 4.17.11 to 4.17.13 [`#8`](https://github.com/audreylopes/readme-generator/pull/8)
+- Develop [`#7`](https://github.com/audreylopes/readme-generator/pull/7)
+- Templating [`#6`](https://github.com/audreylopes/readme-generator/pull/6)
+- Develop [`#5`](https://github.com/audreylopes/readme-generator/pull/5)
 
 ### Commits
 
-- update node version to 10.16 (lts) and upgrade all dependencies [`0966e6d`](https://github.com/ibbatta/readme-generator/commit/0966e6d961c499422eade1277bde09650cabe083)
-- update example readme [`c3ab7bf`](https://github.com/ibbatta/readme-generator/commit/c3ab7bf6756084e2a9e7c5ad6845846ba2a46803)
-
-## v0.1.1 - 2019-08-27
-
-### Merged
-
-- Packagejson fallback [`#35`](https://github.com/ibbatta/readme-generator/pull/35)
-- Jest test [`#34`](https://github.com/ibbatta/readme-generator/pull/34)
-- Change babel polyfill [`#32`](https://github.com/ibbatta/readme-generator/pull/32)
-- Added link to generated example in README [`#30`](https://github.com/ibbatta/readme-generator/pull/30)
-- Develop [`#28`](https://github.com/ibbatta/readme-generator/pull/28)
-- Flag options [`#27`](https://github.com/ibbatta/readme-generator/pull/27)
-- Webpack bundler [`#26`](https://github.com/ibbatta/readme-generator/pull/26)
-- Npm packaging [`#25`](https://github.com/ibbatta/readme-generator/pull/25)
-- Develop [`#24`](https://github.com/ibbatta/readme-generator/pull/24)
-- parse yaml file (FUNDING.yml) to autocomplete patreon questions [`#23`](https://github.com/ibbatta/readme-generator/pull/23)
-- Develop [`#19`](https://github.com/ibbatta/readme-generator/pull/19)
-- Update readme template [`#18`](https://github.com/ibbatta/readme-generator/pull/18)
-- Develop [`#17`](https://github.com/ibbatta/readme-generator/pull/17)
-- Align develop with master [`#14`](https://github.com/ibbatta/readme-generator/pull/14)
-- Refactor settings [`#13`](https://github.com/ibbatta/readme-generator/pull/13)
-- Bump lodash from 4.17.11 to 4.17.13 [`#8`](https://github.com/ibbatta/readme-generator/pull/8)
-- Develop [`#7`](https://github.com/ibbatta/readme-generator/pull/7)
-- Templating [`#6`](https://github.com/ibbatta/readme-generator/pull/6)
-- Develop [`#5`](https://github.com/ibbatta/readme-generator/pull/5)
-
-### Commits
-
-- switch to npm [`08b4509`](https://github.com/ibbatta/readme-generator/commit/08b450974bad10a5a3fd9f4b2a1e9f8a45561bb5)
-- update repository base [`41485e9`](https://github.com/ibbatta/readme-generator/commit/41485e9bf704f3c22ea8c6a844da4c4676f13adb)
-- init repository [`ab38f6c`](https://github.com/ibbatta/readme-generator/commit/ab38f6cffd9f1f03efe03c58b25efd489b4530df)
-- update readme [`4919fb0`](https://github.com/ibbatta/readme-generator/commit/4919fb0f6205d7aa3f903bea7cb3938ebd876587)
-- clean dependencies [`f38323b`](https://github.com/ibbatta/readme-generator/commit/f38323b809db4e8b6fb29a59c529e0ef6b677816)
-- update husky npm config [`10bc250`](https://github.com/ibbatta/readme-generator/commit/10bc250ff2cbef9b5cf61e10f5aae867f40b9e9c)
-- folder refactory and add yaml parser [`a3dc5bc`](https://github.com/ibbatta/readme-generator/commit/a3dc5bcd3637057e819ee4cee37db850e8ce825f)
-- refactor code and inject question dinamically [`eca6d32`](https://github.com/ibbatta/readme-generator/commit/eca6d323d4efc40dc57d1f742f79a790b86b7780)
-- update configurations [`17cf36f`](https://github.com/ibbatta/readme-generator/commit/17cf36ffa66bef859cf62d8bfa5d961f781e9746)
-- megarefactory for settings, utilities and configuration [`edce881`](https://github.com/ibbatta/readme-generator/commit/edce88126b2077e1cea9167d737cb59f742711eb)
+- switch to npm [`08b4509`](https://github.com/audreylopes/readme-generator/commit/08b450974bad10a5a3fd9f4b2a1e9f8a45561bb5)
+- clean dependencies [`f38323b`](https://github.com/audreylopes/readme-generator/commit/f38323b809db4e8b6fb29a59c529e0ef6b677816)
+- update husky npm config [`10bc250`](https://github.com/audreylopes/readme-generator/commit/10bc250ff2cbef9b5cf61e10f5aae867f40b9e9c)
+- refactor code and inject question dinamically [`eca6d32`](https://github.com/audreylopes/readme-generator/commit/eca6d323d4efc40dc57d1f742f79a790b86b7780)
+- minor [`6543b64`](https://github.com/audreylopes/readme-generator/commit/6543b644aaea4cd0bfbb2f1ec8c8873004b42f93)
+- fix [`4af7ad6`](https://github.com/audreylopes/readme-generator/commit/4af7ad61337b5537dcf5bebff8b1545e5bc4aea2)
+- refactory and add tests [`4936dca`](https://github.com/audreylopes/readme-generator/commit/4936dcad4c9d757806b1730c1ce077dc93a47b03)
+- create partialhbs reader [`8d3e6da`](https://github.com/audreylopes/readme-generator/commit/8d3e6da223fe133e0db291de444f32002e27e754)
+- update husky npm config [`009f522`](https://github.com/audreylopes/readme-generator/commit/009f52250a8f3a449587e48ecb2c5eaa5dc87405)
+- split code and add more test [`60e6a5b`](https://github.com/audreylopes/readme-generator/commit/60e6a5b179e48c2be2b654fe6601dfa2ed6ec481)
