+++ test/settings/file.settings.test.js
@@ -1,4 +1,4 @@
-import fileSettings from './file.settings';
+import fileSettings from '../../src/settings/file.settings';
 
 test('check if fileSettings exists and is not empty', () => {
   expect(fileSettings).not.toBeNull();
