+++ test/utilities/handlebar.utils.test.js
@@ -1,4 +1,4 @@
-import hbsUtils from './handlebar.utils';
+import hbsUtils from '../../src/utilities/handlebar.utils';
 
 test('check if fileUtils exists and is not empty', () => {
   expect(hbsUtils).not.toBeNull();
