+++ test/utilities/question.utils.test.js
@@ -1,4 +1,4 @@
-import questionUtils from './question.utils';
+import questionUtils from '../../src/utilities/question.utils';
 
 test('check if fileUtils exists and is not empty', () => {
   expect(questionUtils).not.toBeNull();
