+++ test/utilities/ststem.utils.test.js
@@ -1,4 +1,4 @@
-import systemUtils from './system.utils';
+import systemUtils from '../../src/utilities/system.utils';
 
 test('check if fileUtils exists and is not empty', () => {
   expect(systemUtils).not.toBeNull();
